Move legacy config cleanup out of config api
