+++ RebornCore/src/main/java/reborncore/common/config/Config.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of RebornCore, licensed under the MIT License (MIT).
  *
- * Copyright (c) 2021 TeamReborn
+ * Copyright (c) 2016-2017 TeamReborn
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,42 +24,13 @@
 
 package reborncore.common.config;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import org.jetbrains.annotations.ApiStatus;
 
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface Config {
-
-	/**
-	 * This the category of the config
-	 *
-	 * @return {@link String}
-	 */
-	String category() default "config";
-
-	/**
-	 * This is the key for the config, the default is the field name.
-	 *
-	 * @return {@link String}
-	 */
-	String key() default "";
-
-	/**
-	 * This is a comment that will be supplied along with the config, use this to explain what the config does
-	 *
-	 * @return {@link String}
-	 */
-	String comment() default "";
-
-	/**
-	 * This is the config file name, the default is {@code config.cgf},
-	 * use this if you wish to split the config into more than one file.
-	 *
-	 * @return {@link String}
-	 */
-	String config() default "config";
-
+/**
+ * This is the root of a configuration tree.
+ */
+@ApiStatus.NonExtendable
+public interface Config extends ConfigGroup {
+	@Override
+	Config comment(String comment);
 }
