+++ RebornCore/src/main/java/reborncore/common/config/ConfigGroup.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package reborncore.common.config2;
+package reborncore.common.config;
 
 import java.util.List;
 import java.util.function.Supplier;
