--- RebornCore/src/main/java/reborncore/common/config2/impl/ConfigDirWatcher.java
@@ -1,67 +1,0 @@
-/*
- * This file is part of RebornCore, licensed under the MIT License (MIT).
- *
- * Copyright (c) 2016-2017 TeamReborn
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package reborncore.common.config2.impl;
-
-import java.io.IOException;
-import java.nio.file.FileSystems;
-import java.nio.file.Path;
-import java.nio.file.StandardWatchEventKinds;
-import java.nio.file.WatchEvent;
-import java.nio.file.WatchKey;
-import java.nio.file.WatchService;
-
-import net.fabricmc.loader.api.FabricLoader;
-
-public class ConfigDirWatcher {
-	private static ConfigDirWatcher INSTANCE = null;
-
-	public static void startWatching() throws IOException {
-		if (INSTANCE == null) {
-			INSTANCE = new ConfigDirWatcher();
-		}
-	}
-
-	private final Path configDir;
-	private final WatchService watchService;
-
-	private ConfigDirWatcher() throws IOException {
-		configDir = FabricLoader.getInstance().getConfigDir();
-		watchService = FileSystems.getDefault().newWatchService();
-
-		configDir.register(
-				watchService,
-				StandardWatchEventKinds.ENTRY_MODIFY);
-	}
-
-	private void poll() throws InterruptedException {
-		WatchKey key;
-		while ((key = watchService.take()) != null) {
-			for (WatchEvent<?> event : key.pollEvents()) {
-				// TODO: Handle the event
-			}
-			key.reset();
-		}
-	}
-}
